requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
httpx==0.23.3
python-dotenv==1.0.0
requests==2.31.0
supabase==1.0.3
Pillow==10.1.0 